requirements.txt
asgiref==3.5.0
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.5.0
Django==3.2
django-environ==0.8.1
fonttools==4.31.2
gunicorn==20.1.0
html5lib==1.1
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.0.0
zopfli==0.2.1
